hw1rg

+++ FVCalc.java
@@ -1,7 +1,7 @@
 // Computes the future value of a saving investment.
 public class FVCalc {
 	public static void main(String[] args){
-		double c= Double.parseDouble(args[0]);
+		int c= Integer.parseInt(args[0]);
 		double r= Double.parseDouble(args[1]);
 		int n= Integer.parseInt(args[2]); 
        double m=(r/100)+1;
